MLK-15318-4 rework the build system
Currently there is no common infrastructure to build the test and
because of that there is a lot of duplication, and no consistency on
how test are built. Also, because files are moved around, incremental
builds don't work properly. Not the least, build errors are skipped and
there is too much info output during the build, which makes it hard to
understand what is going on.

This patch completely reworks the build system to fix all of the
described issues. It also removes the doc target from the default all
target list, as it does not build succesfully, to avoid breaking
default builds.

To avoid code duplication each sub-makefile only needs to define the
following information:

BUILD - a list of executables to build
COPY - a list of files to copy
DIR - destination directory where the files should be put
CFLAGS - additional flags to be used for compiling
LDFLAGS - additional flags to be used for linking

Rules will be automatically generated to build the executables and
copy the files to the proper location.

For a given executable the source filename is determined as following:
if the executable is named exec.out the source file is expected to be
exec.c, otherwise if the executable is named is exec the source is
expected to be exec.c. For example:

BUILD = a.out b

will use a.c and b.c as sources.

If an executable has multiple sources, it must define a new variable
with the same name as the executable that contains all of the object
names. For example:

BUILD = a.out
a.out = a.o b.o

will built a.out from a.c and b.c.

By default the output of the build is condensed, in a similar fashion
with how the Linux kernel does it. Use V=1 to enable full build
output.

Signed-off-by: Octavian Purdila <octavian.purdila@nxp.com>